修改User中的entity对象

+++ src/main/java/com/hospital/dao/UserDao.java
@@ -5,5 +5,5 @@ import com.hospital.entity.User;
 /**
  * Created by Jimmy on 2017/3/11.
  */
-public interface UserDao extends DomainDao<User,Long> {
+public interface UserDao extends DomainDao<User,Integer> {
 }
