+++ src/main/java/com/hospital/service/UserService.java
@@ -4,5 +4,5 @@ package com.hospital.service;
  * Created by Jimmy on 2017/3/11.
  */
 public interface UserService {
-    Long saveUser();
+    Integer saveUser();
 }
